% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_panel.R
\name{run_panel}
\alias{run_panel}
\title{Sensory Board Panel}
\usage{
run_panel(
  products_file,
  attributes_file,
  answers_dir = "Answers",
  dest_url = NULL,
  numeric_range = c(0, 10)
)
}
\arguments{
\item{products_file}{A character path to the csv file containing the products to evaluate.}

\item{attributes_file}{A character path to the csv file containing the attributes to evaluate.}

\item{answers_dir}{A character path to the folder in which to save user responses (if it does not
exist, it will create it).}

\item{dest_url}{An optional character including the URL to use as destination host and port.
For example: 192.168.100.7:4000 .}

\item{numeric_range}{A numeric vector indicating the range for numeric inputs.}
}
\description{
Sensory Board Panel
}
